Combine Validate and Translate functions
